+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [0.1.90] - 2017-11-05
+### Changed
+- New features in the `Address.geolocate()` method.
+### Added
+- The `Address.normalize()` method.
+- The `session list interview` and `dispatch interview` configuration
+  directives.
+
 ## [0.1.89] - 2017-11-04
 ### Added
 - Ability to use `json=1` with `/user/sign-in`, `/list`, and